+++ index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import 'react-photo-view/dist/react-photo-view.css';
+
 import App from './App';
 import { AppProvider } from './store/AppContext';